search bar + added webpage for when nothing is found at the api call
added a simple search bar on the home page that lets the user search for a cocktail, which makes an api call and displays cocktail image and ingredients. also, made a webpage to call when an api call returns an empty json file so that the website doesnt crash in this case.